TL header clarifications
This patch adds further details to clarify the meaning of some fields in
the TL header. The checksum calculation calls out that the checksum
field itself must be assumed as 0 for purposes of calculating the
checksum. The size field clarifies that it includes the inter-TE padding
behind the last TE (and must therefore be a multiple of 16 that can be
directly used to calculate the start offset for the next TE to be
added). The version field clarifies the contract created by versions
(backwards-compatibility for reading only).

Signed-off-by: Julius Werner <jwerner@chromium.org>