Change TE header size to 8 bytes and introduce alignment field
This patch changes the TE header format to reduce the size of some
fields and remove reserved space, making the new header 8 bytes long. It
also updates the documentation for the size calculation to more
explicitly refer to the fields in the header. The alignment requirement
for TE headers and the whole TL is similarly reduced to 8. In order to
still support TEs with larger alignment requirement, a new *alignment*
field is added to the TL header to track the maximum required alignment
of any TE, and allow relocation of the whole TL while preserving those
alignment requirements. The XFERLIST_VOID entry description is updated
to also be used as padding for aligning later TEs, and to explicitly
allow it to be overwritten with other TEs.

Signed-off-by: Julius Werner <jwerner@chromium.org>