Clarify expectations for tag allocation process
This patch rewrites the section about entry type ranges to try to make
it more clear how tags are supposed to be allocated and what
standardized tags can be used for, and how non-standardized tags should
be avoided to preserve interoperability. Also move the non-standardized
range to the end of the tag space, as discussed during review.

Signed-off-by: Julius Werner <jwerner@chromium.org>