Format field names as :code: instead of emphasis
This patch changes the formatting of all field names (e.g. tag_id) in
the spec from emphasis (italics) to :code: (monospace). This looks
better and is more readable particularly when writing more complicated
calculations involving multiple fields. Since this formatting is
required so often, set it as the default-role.

Signed-off-by: Julius Werner <jwerner@chromium.org>