Fix relocation algorithm description
This patch fixes a mistake that crept in during review of PR #6 (patch
062d542). The `alignment_offset` calculated from `tl_base_addr` was not
superfluous in that calculation -- in fact, this is the only way in
which `tl_base_addr` entered the calculation at all. If we don't have
this, relocation for base addresses that do not already happen to match
the highest alignment requirement won't work.

Also fix another small mistake in the same algorithm regarding
calculating the new `tl.max_size`.

Signed-off-by: Julius Werner <jwerner@chromium.org>