Account for TE header in algorithm to add an aligned TE
For TEs with special alignment requirement, the point is to align the
data portion, not the TE header. This needs to be taken into account
when calculating how much padding needs to be inserted to reach the
correct alignment. This patch fixes the algorithm description for adding
aligned TEs to account for that.

Signed-off-by: Julius Werner <jwerner@chromium.org>