Adjust terminoloy for size
The word 'size' is pretty generic and appears through the spec. It
refers to the about of space currently used by the TEs, so rename it to
used_size

The term `max_size` is a little confusing since it is really the
current size of the TL including its spare space. Rename this to
total_size indicating that is the total of all TEs and space space.

It also happens to match the devicetree terminology.

Fixes: #22

Signed-off-by: Simon Glass <sjg@chromium.org>