Fix broken link to references section
The references section doesn't appear in HTML builds because the page
wasn't titled. Add a title and fix the order it appears in the table of
contents. Also make sure the sections are numbered, without this numrefs
that work in PDF builds don't work in html builds.

Because of some intracacies in LaTex builders that cause empty
reference pages to be rendered [1], a workaround using the raw directive
is applied to remove the phantom page.

[1] https://github.com/sphinx-doc/sphinx/issues/4775

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>