Add FF-A SP binary TE type
Allocate a TE type in the Trusted Firmware range for FF-A
SP binaries.

Hafnium SPMC supports the SP's artefacts to be packaged as a
Transfer List generated at build time (with TF-A's tlc tool).
The respective Transfer List contents: SP binary/image, the FF-A
manifest, and other optional resoures (e.g. HOB).

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I34fada0dca2b5d6f30fcf2b7a54d93028aa59727