Add an FF-A manifest TE type
Specify an FF-A manifest TE type to the trusted firmware
range.

This entry is to be used in the TL package containing
an SP's artefacts, created at build time using the
tlc tool from TF-A, and consumed by an SPMC implementation.

For the moment, Hafnium as the S-EL2 SPMC is used.

This patchset resolves issue #51:
https://github.com/FirmwareHandoff/firmware_handoff/issues/51

Signed-off-by: J-Alves <joao.alves@arm.com>