Relax language around use of non-standardized tag id range
Slightly relax the language around the use of the non-standardized tag id
range, to allow the use between closely coupled components that must only be
used within a particular firmware implementation. In some cases, the TE
definition is continuously evolving or very complex to define, which makes
standardization an onerous requirement.

Signed-off-by: Dan Handley <dan.handley@arm.com>